// pork-ionic-angular/src/app/interfaces/user-interface.ts
export class UserInterface {
    nombre: string ;
    apellido: string;
    celular: number;
    granja: string;
    localizacion: string;
    email: string;
    password: string;
    id: any = '';
    codigoMostrar: string = '';
    estado: boolean = false;
    idUsuario: string = '';
    pushUserId: string = '';
}